changed layout of codemodel and added conditionalschemas
