+++ codemodel/.resources/model/json/types.json
@@ -32,6 +32,7 @@
         "byte-array",
         "char",
         "choice",
+        "conditional",
         "credential",
         "date",
         "date-time",
@@ -40,6 +41,7 @@
         "integer",
         "number",
         "sealed-choice",
+        "sealed-conditional",
         "string",
         "unixtime",
         "uri",
@@ -65,6 +67,7 @@
         "byte-array",
         "char",
         "choice",
+        "conditional",
         "constant",
         "credential",
         "date",
@@ -80,6 +83,7 @@
         "odata-query",
         "or",
         "sealed-choice",
+        "sealed-conditional",
         "string",
         "unixtime",
         "uri",
