+++ codemodel/.resources/model/yaml/master.yaml
@@ -53,6 +53,115 @@ definitions:
   CSharpLanguage:
     type: object
     additionalProperties: false
+  ConditionalType:
+    type: object
+    additionalProperties: false
+    properties:
+      type:
+        description: the schema type
+        $ref: './types.yaml#/definitions/AllSchemaTypes'
+      apiVersions:
+        type: array
+        description: API versions that this applies to. Undefined means all versions
+        items:
+          $ref: './master.yaml#/definitions/ApiVersion'
+      defaultValue:
+        description: 'If the value isn''t sent on the wire, the service will assume this'
+      deprecated:
+        description: 'deprecation information -- ie, when this aspect doesn''t apply and why'
+        $ref: './master.yaml#/definitions/Deprecation'
+      example:
+        description: example information
+      externalDocs:
+        description: External Documentation Links
+        $ref: './master.yaml#/definitions/ExternalDocumentation'
+      serialization:
+        description: per-serialization information for this Schema
+        $ref: './master.yaml#/definitions/SerializationFormats'
+      summary:
+        type: string
+        description: a short description
+      extensions:
+        description: additional metadata extensions dictionary
+        $ref: './master.yaml#/definitions/Dictionary<any>'
+      language:
+        description: per-language information for Schema
+        $ref: './master.yaml#/definitions/Languages'
+      protocol:
+        description: per-protocol information for this aspect
+        $ref: './master.yaml#/definitions/Protocols'
+    required:
+      - language
+      - protocol
+      - type
+  ConditionalType_1:
+    type: object
+    additionalProperties: false
+    properties:
+      type:
+        description: the schema type
+        $ref: './types.yaml#/definitions/AllSchemaTypes'
+      apiVersions:
+        type: array
+        description: API versions that this applies to. Undefined means all versions
+        items:
+          $ref: './master.yaml#/definitions/ApiVersion'
+      defaultValue:
+        description: 'If the value isn''t sent on the wire, the service will assume this'
+      deprecated:
+        description: 'deprecation information -- ie, when this aspect doesn''t apply and why'
+        $ref: './master.yaml#/definitions/Deprecation'
+      example:
+        description: example information
+      externalDocs:
+        description: External Documentation Links
+        $ref: './master.yaml#/definitions/ExternalDocumentation'
+      serialization:
+        description: per-serialization information for this Schema
+        $ref: './master.yaml#/definitions/SerializationFormats'
+      summary:
+        type: string
+        description: a short description
+      extensions:
+        description: additional metadata extensions dictionary
+        $ref: './master.yaml#/definitions/Dictionary<any>'
+      language:
+        description: per-language information for Schema
+        $ref: './master.yaml#/definitions/Languages'
+      protocol:
+        description: per-protocol information for this aspect
+        $ref: './master.yaml#/definitions/Protocols'
+    required:
+      - language
+      - protocol
+      - type
+  ConditionalValue:
+    type: object
+    description: an individual value in a ConditionalSchema
+    additionalProperties: false
+    properties:
+      source:
+        type:
+          - string
+          - number
+          - boolean
+        description: the actual value
+      target:
+        type:
+          - string
+          - number
+          - boolean
+        description: the actual value
+      extensions:
+        description: additional metadata extensions dictionary
+        $ref: './master.yaml#/definitions/Dictionary<any>'
+      language:
+        description: per-language information for this value
+        $ref: './master.yaml#/definitions/Languages'
+    required:
+      - language
+      - source
+      - target
   Contact:
     type: object
     description: contact information
@@ -280,10 +389,6 @@ definitions:
     description: 'represents a single callable endpoint with a discrete set of inputs, and any number of output possibilities (responses or exceptions)'
     additionalProperties: false
     properties:
-      $key:
-        type: string
-      description:
-        type: string
       apiVersions:
         type: array
         description: API versions that this applies to. Undefined means all versions
@@ -324,8 +429,6 @@ definitions:
         description: per-protocol information for this aspect
         $ref: './master.yaml#/definitions/Protocols'
     required:
-      - $key
-      - description
       - language
       - protocol
       - request
@@ -358,8 +461,6 @@ definitions:
         description: suggested implementation location for this parameter
         $ref: './enums.yaml#/definitions/ImplementationLocation'
     required:
-      - $key
-      - description
       - language
       - protocol
       - schema
@@ -380,8 +481,6 @@ definitions:
         type: string
         description: the wire name of this property
     required:
-      - $key
-      - description
       - language
       - protocol
       - schema
@@ -480,18 +579,18 @@ definitions:
     description: 'common base interface for properties, parameters and the like.'
     additionalProperties: false
     properties:
-      $key:
-        type: string
       schema:
         description: the schema of this Value
         $ref: './schemas.yaml#/definitions/Schema'
-      description:
-        type: string
       apiVersions:
         type: array
         description: API versions that this applies to. Undefined means all versions
         items:
           $ref: './master.yaml#/definitions/ApiVersion'
+      assumedValue:
+        description: the value that the remote will assume if this value is not present
+      clientDefaultValue:
+        description: the value that the client should provide if the consumer doesn't provide one
       deprecated:
         description: 'deprecation information -- ie, when this aspect doesn''t apply and why'
         $ref: './master.yaml#/definitions/Deprecation'
@@ -517,8 +616,6 @@ definitions:
         description: per-protocol information for this aspect
         $ref: './master.yaml#/definitions/Protocols'
     required:
-      - $key
-      - description
       - language
       - protocol
       - schema
