+++ codemodel/exports.ts
@@ -26,7 +26,21 @@ export * from './model/http/HttpMethod';
 export * from './model/http/ParameterLocation';
 export * from './model/http/security';
 export * from './model/http/SerializationStyle';
-export * from './model/http/server';
+
 export * from './model/http/status-code';
 export * from './model/yaml-schema';
 export * from './tag';
+export * from './model/common/schemas/any';
+export * from './model/common/schemas/array';
+export * from './model/common/schemas/binary';
+export * from './model/common/schemas/choice';
+export * from './model/common/schemas/conditional';
+export * from './model/common/schemas/constant';
+export * from './model/common/schemas/dictionary';
+export * from './model/common/schemas/flag';
+export * from './model/common/schemas/number';
+export * from './model/common/schemas/object';
+export * from './model/common/schemas/primitive';
+export * from './model/common/schemas/relationship';
+export * from './model/common/schemas/string';
+export * from './model/common/schemas/time';
